Cleanup and fixes for Foot Reverse code
I have split the code into three files.  The main code, the config for
drive and config for the MarcDuino.  This significantly cleans up the
main .ino file, making it much easier to read and to configure the
relevant sections.

Bug fixes for the reverse foot code have been implemented.  The
reverse the motor direction of the relevant motor.

The Servo code has been updated to output a microsecond value rather
than an angle.  When sending the angle to the SparkMax controller,
things worked as expected in "forward" direction, but in reverse the
motor would ramp to max speed.  (Not ideal!)  This is resolved by sending
a microsecond value.  For simplicity the code still calculates
the values in angle, but then converts them to microsecond to send to the
controller.  This shouldn't affect any other controller, as this is the
true PWM signal anyway!